requirements.txt
pillow
requests
filestack-python
google-generativeai
gunicorn 
pytelegrambotapi
